+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Security:
 
 Changes:
 
+* `RubyTruffleError` has been replaced with `RuntimeError`.
+
 * C++ libraries like `libc++` are now not needed if you don't run C++
   extensions. `libc++abi` is now never needed. Documentation updated to make it
   more clear what the minimum requirements for pure Ruby, C extensions, and C++